[Store][ChromaDb] Split store as dedicated package
