+++ .github/workflows/validation.yaml
@@ -1,0 +1,81 @@
+name: Validation
+
+on:
+    push:
+        paths:
+            - 'src/store/src/Bridge/**/composer.json'
+            - 'src/ai-bundle/config/options.php'
+            - '.github/workflows/validation.yaml'
+    pull_request:
+        paths:
+            - 'src/store/src/Bridge/**/composer.json'
+            - 'src/ai-bundle/config/options.php'
+            - '.github/workflows/validation.yaml'
+
+concurrency:
+    group: ${{ github.workflow }}-${{ github.head_ref || github.run_id }}
+    cancel-in-progress: true
+
+jobs:
+    validate_stores:
+        name: Validate Store Bridge Naming
+        runs-on: ubuntu-latest
+        steps:
+            - name: Checkout
+              uses: actions/checkout@v5
+
+            - name: Validate store bridge naming conventions
+              run: |
+                  #!/bin/bash
+                  set -e
+
+                  ERRORS=0
+
+                  # Find all store bridges with composer.json
+                  for composer_file in src/store/src/Bridge/*/composer.json; do
+                      if [[ ! -f "$composer_file" ]]; then
+                          continue
+                      fi
+
+                      # Get the bridge directory name (e.g., ChromaDb)
+                      bridge_dir=$(dirname "$composer_file")
+                      bridge_name=$(basename "$bridge_dir")
+
+                      # Get the package name from composer.json
+                      package_name=$(jq -r '.name' "$composer_file")
+
+                      # Expected package name format: symfony/ai-{lowercase-with-dashes}-store
+                      # Convert PascalCase to kebab-case (e.g., ChromaDb -> chroma-db)
+                      expected_kebab=$(echo "$bridge_name" | sed 's/\([a-z]\)\([A-Z]\)/\1-\2/g' | tr '[:upper:]' '[:lower:]')
+                      expected_package="symfony/ai-${expected_kebab}-store"
+
+                      if [[ "$package_name" != "$expected_package" ]]; then
+                          echo "::error file=$composer_file::Package name '$package_name' does not match expected '$expected_package' for bridge '$bridge_name'"
+                          ERRORS=$((ERRORS + 1))
+                      else
+                          echo "✓ $bridge_name: package name '$package_name' is correct"
+                      fi
+
+                      # Check options.php for the config key (should be lowercase without dashes/underscores)
+                      expected_config_key=$(echo "$bridge_name" | tr '[:upper:]' '[:lower:]')
+                      options_file="src/ai-bundle/config/options.php"
+
+                      if [[ -f "$options_file" ]]; then
+                          # Look for ->arrayNode('configkey') under the store section
+                          if ! grep -q -- "->arrayNode('$expected_config_key')" "$options_file"; then
+                              echo "::error file=$options_file::Missing or incorrect config key for bridge '$bridge_name'. Expected '->arrayNode('$expected_config_key')' in store configuration"
+                              ERRORS=$((ERRORS + 1))
+                          else
+                              echo "✓ $bridge_name: config key '$expected_config_key' found in options.php"
+                          fi
+                      fi
+                  done
+
+                  if [[ $ERRORS -gt 0 ]]; then
+                      echo ""
+                      echo "::error::Found $ERRORS naming convention violation(s)"
+                      exit 1
+                  fi
+
+                  echo ""
+                  echo "All store bridge naming conventions are valid!"
