+++ examples/composer.json
@@ -26,7 +26,7 @@
         "ext-pdo": "*",
         "ext-redis": "*",
         "async-aws/bedrock-runtime": "^1.1",
-        "codewithkyrian/chromadb-php": "^0.4.0",
+        "symfony/ai-chroma-db-store": "@dev",
         "codewithkyrian/transformers": "^0.6.2",
         "doctrine/dbal": "^3.3|^4.0",
         "google/auth": "^1.47",
