+++ src/store/src/Bridge/ChromaDb/README.md
@@ -1,0 +1,12 @@
+ChromaDB Store
+==============
+
+Provides [ChromaDB](https://www.trychroma.com/) vector store integration for Symfony AI Store.
+
+Resources
+---------
+
+ * [Contributing](https://symfony.com/doc/current/contributing/index.html)
+ * [Report issues](https://github.com/symfony/ai/issues) and
+   [send Pull Requests](https://github.com/symfony/ai/pulls)
+   in the [main Symfony AI repository](https://github.com/symfony/ai)
